dvipos: added ChangeLog, revised man page, fixed version strings
Consistent with the GPL's section 2(a), I added a ChangeLog and
updated the version strings to be 20070107.  People will then know
that they are running a different dvipos from the CTAN -20030628
version (which didn't work with current versions of ConTeXt), and can
blame the right people (i.e. me) if one of my two changes to dvicore.c
causes a problem.  No binaries need rebuilding because all the changes
are to uncompiled files (ChangeLog, dvipos.1) or to uncompiled parts
of files (comments).  In utils.h, the #define VERSION was already
20070107, so it remains unchanged.



git-svn-id: svn://tug.org/texlive/trunk/Build/source@3556 c570f23f-e606-0410-a88d-b1316a301751