change the condition for the "special texlive build behavior": previously it was triggered just by the --enable-tetex-build configure option, but some distributions are using both --enable-tetex-build and at the same time redefine the datadir, docdir, mandir, etc.
now, the "special texlive build behavior" is enabled when both
the --enable-tetex-build option is used, AND $datadir is the same
as $prefix, which is the case when using the default Build script.

the "special texlive build behavior" for xindy consists of:

* installing "xindy" and "texindy" as symlinks pointing to
  scripts in the texmf/scripts/xindy/ directory, rather than
  copying them as regular files into bindir.

* installing xindy.mem into bindir (rather than into pkglibdir),
  as in texlive, it is the only platform-dependent directory, and
  there's no per-platform lib directory.

* redefining datadir, docdir and mandir to the layout used in texlive:
  datadir="$prefix/texmf"
  docdir="$prefix/texmf/doc/$PACKAGE_NAME"
  mandir="$prefix/texmf/doc/man"

* disabling docs generation and make-rules build (because they
  require working latex and pdflatex during build). in normal
  texlive builds, the docs and make-rules are installed in the
  texmf tree, and need not be re-built.

if distributor wants to build the whole texlive, with xindy, and
uses the --enable-tetex-build, but redefines datadir, we disable
the above "special texlive build behavior".

in this case, however, please keep in mind that you need the
working latex (for make-rules) and pdflatex (for docs) during
xindy build. if you do not have it, add --disable-make-rules and
--disable-docs configure options, but keep in mind that without
the make-rules, xindy's install is incomplete, and you need to
use the files supplied by texlive's Master/texmf/xindy tree.


git-svn-id: svn://tug.org/texlive/trunk/Build/source@8199 c570f23f-e606-0410-a88d-b1316a301751