change to the Build script in r8412 (which allows the Build to work on systems such as Solaris which don't have BSD-compatible "install" program) made ZZIPLIBSRCDIR an absolute directory, so ../$(ZZIPLIBSRCDIR)/zzip/Makefile won't work.
the rules for zziplib are broken, in that they assume that the
Work directory is a sub-directory of the source directory, and
use specific relative paths, and/or they don't work for absolute
directories.

now, removing changing ../$(ZZIPLIBSRCDIR)/ to $(ZZIPLIBSRCDIR)/
will rely on $(ZZIPLIBSRCDIR) being an absolute directory.  so if
someone manually calls configure (rather than via the Build
script) he needs to run it with absolute path, e.g. even
`pwd`/../configure rather than ../configure


git-svn-id: svn://tug.org/texlive/trunk/Build/source@8417 c570f23f-e606-0410-a88d-b1316a301751